Fix "game" / "name" discrepancy
The Game class in code, which is deserialized from JSON, had a property
called "name". However, the actual JSON data called that property
"game". This change updates the JSON to call the property "name", but
also adds a fix to the code so that older JSON that calls it "game" will
still load and work correctly.